requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2.15
django-blacklist==0.6.1
django-cors-headers==3.13.0
django-filter==22.1
django-ratelimit==3.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
